Added new versions.
Changed quite a bit. Now the ball is fired by using spacebar and no longer follows the user's mouse. Support has been added for changing angle/velocity for the next ball to be fired. Tried fixing a list of options for changing level of realism on the fly too, but I couldn't figure out how to bind the signal correctly.

Some things to consider adding (not that we have much time left to do it, haha):
Changing radius and mass of ball on the fly.
Uh.. I guess that's pretty much it?